+++ src/py_balance.cc
@@ -81,11 +81,11 @@ namespace {
 #endif
 
   long balance_len(balance_t& bal) {
-    return bal.amounts.size();
+    return static_cast<long>(bal.amounts.size());
   }
 
   amount_t balance_getitem(balance_t& bal, long i) {
-    long len = bal.amounts.size();
+    long len = static_cast<long>(bal.amounts.size());
 
     if (labs(i) >= len) {
       PyErr_SetString(PyExc_IndexError, _("Index out of range"));
