+++ src/py_xact.cc
@@ -44,7 +44,7 @@ namespace {
 
   long posts_len(xact_base_t& xact)
   {
-    return xact.posts.size();
+    return static_cast<long>(xact.posts.size());
   }
 
   post_t& posts_getitem(xact_base_t& xact, long i)
@@ -53,7 +53,7 @@ namespace {
     static xact_base_t * last_xact = NULL;
     static posts_list::iterator elem;
 
-    long len = xact.posts.size();
+    long len = static_cast<long>(xact.posts.size());
 
     if (labs(i) >= len) {
       PyErr_SetString(PyExc_IndexError, _("Index out of range"));
